--- easycv/models/detection/dab_detr/__init__.py
@@ -1,2 +1,0 @@
-from .dab_detr_head import DABDETRHead
-from .dab_detr_transformer import DABDetrTransformer
