--- easycv/models/detection/dab_detr/dab_detr_transformer.py
@@ -1,529 +1,0 @@
-# Copyright (c) 2022 IDEA. All Rights Reserved.
-# Copyright (c) Alibaba, Inc. and its affiliates.
-"""
-DETR Transformer class.
-Copy-paste from torch.nn.Transformer with modifications:
-    * positional encodings are passed in MHattention
-    * extra LN at the end of encoder is removed
-    * decoder returns a stack of activations from all decoding layers
-"""
-import math
-from typing import Optional
-
-import torch
-import torch.nn.functional as F
-from torch import Tensor, nn
-
-from easycv.models.builder import NECKS
-from easycv.models.detection.utils import (MLP, TransformerEncoder,
-                                           TransformerEncoderLayer,
-                                           _get_activation_fn, _get_clones)
-from .attention import MultiheadAttention
-
-
-def inverse_sigmoid(x, eps=1e-3):
-    x = x.clamp(min=0, max=1)
-    x1 = x.clamp(min=eps)
-    x2 = (1 - x).clamp(min=eps)
-    return torch.log(x1 / x2)
-
-
-@NECKS.register_module
-class DABDetrTransformer(nn.Module):
-
-    def __init__(self,
-                 in_channels=1024,
-                 query_dim=4,
-                 d_model=512,
-                 nhead=8,
-                 num_encoder_layers=6,
-                 num_decoder_layers=6,
-                 dim_feedforward=2048,
-                 dropout=0.1,
-                 activation='relu',
-                 normalize_before=False,
-                 return_intermediate_dec=False,
-                 keep_query_pos=False,
-                 query_scale_type='cond_elewise',
-                 modulate_hw_attn=True,
-                 bbox_embed_diff_each_layer=False,
-                 temperatureH=20,
-                 temperatureW=20):
-        super().__init__()
-        assert query_scale_type in [
-            'cond_elewise', 'cond_scalar', 'fix_elewise'
-        ]
-
-        self.input_proj = nn.Conv2d(in_channels, d_model, kernel_size=1)
-        self.positional_encoding = PositionEmbeddingSineHW(
-            d_model // 2,
-            temperatureH=temperatureH,
-            temperatureW=temperatureW,
-            normalize=True)
-
-        encoder_layer = TransformerEncoderLayer(d_model, nhead,
-                                                dim_feedforward, dropout,
-                                                activation, normalize_before)
-        encoder_norm = nn.LayerNorm(d_model) if normalize_before else None
-        self.encoder = TransformerEncoder(
-            encoder_layer,
-            num_encoder_layers,
-            encoder_norm,
-            query_scale_type=query_scale_type)
-
-        decoder_layer = TransformerDecoderLayer(
-            d_model,
-            nhead,
-            dim_feedforward,
-            dropout,
-            activation,
-            normalize_before,
-            keep_query_pos=keep_query_pos)
-        decoder_norm = nn.LayerNorm(d_model)
-        self.decoder = TransformerDecoder(
-            decoder_layer,
-            num_decoder_layers,
-            decoder_norm,
-            return_intermediate=return_intermediate_dec,
-            d_model=d_model,
-            query_dim=query_dim,
-            keep_query_pos=keep_query_pos,
-            query_scale_type=query_scale_type,
-            modulate_hw_attn=modulate_hw_attn,
-            bbox_embed_diff_each_layer=bbox_embed_diff_each_layer)
-
-        self.d_model = d_model
-        self.nhead = nhead
-
-    def init_weights(self):
-        for p in self.named_parameters():
-            if 'input_proj' in p[0] or 'positional_encoding' in p[
-                    0] or 'bbox_embed' in p[0]:
-                continue
-            if p[1].dim() > 1:
-                nn.init.xavier_uniform_(p[1])
-
-    def forward(self, src, img_metas, query_embed, tgt, attn_mask=None):
-        src = src[0]
-
-        # construct binary masks which used for the transformer.
-        # NOTE following the official DETR repo, non-zero values representing
-        # ignored positions, while zero values means valid positions.
-        bs = src.size(0)
-        input_img_h, input_img_w = img_metas[0]['batch_input_shape']
-        mask = src.new_ones((bs, input_img_h, input_img_w))
-        for img_id in range(bs):
-            img_h, img_w, _ = img_metas[img_id]['img_shape']
-            mask[img_id, :img_h, :img_w] = 0
-
-        src = self.input_proj(src)
-        # interpolate masks to have the same spatial shape with x
-        mask = F.interpolate(
-            mask.unsqueeze(1), size=src.shape[-2:]).to(torch.bool).squeeze(1)
-        # position encoding
-        pos_embed = self.positional_encoding(mask)  # [bs, embed_dim, h, w]
-        # outs_dec: [nb_dec, bs, num_query, embed_dim]
-
-        # flatten NxCxHxW to HWxNxC
-        src = src.flatten(2).permute(2, 0, 1)
-        pos_embed = pos_embed.flatten(2).permute(2, 0, 1)
-        mask = mask.flatten(1)
-
-        memory = self.encoder(src, src_key_padding_mask=mask, pos=pos_embed)
-
-        hs, references = self.decoder(
-            tgt,
-            memory,
-            tgt_mask=attn_mask,
-            memory_key_padding_mask=mask,
-            pos=pos_embed,
-            refpoints_unsigmoid=query_embed)
-        return hs, references
-
-
-class PositionEmbeddingSineHW(nn.Module):
-    """
-    This is a more standard version of the position embedding, very similar to the one
-    used by the Attention is all you need paper, generalized to work on images.
-    """
-
-    def __init__(self,
-                 num_pos_feats=64,
-                 temperatureH=10000,
-                 temperatureW=10000,
-                 normalize=False,
-                 scale=None):
-        super().__init__()
-        self.num_pos_feats = num_pos_feats
-        self.temperatureH = temperatureH
-        self.temperatureW = temperatureW
-        self.normalize = normalize
-        if scale is not None and normalize is False:
-            raise ValueError('normalize should be True if scale is passed')
-        if scale is None:
-            scale = 2 * math.pi
-        self.scale = scale
-
-    def forward(self, mask):
-        assert mask is not None
-        not_mask = ~mask
-        y_embed = not_mask.cumsum(1, dtype=torch.float32)
-        x_embed = not_mask.cumsum(2, dtype=torch.float32)
-
-        # import ipdb; ipdb.set_trace()
-
-        if self.normalize:
-            eps = 1e-6
-            y_embed = y_embed / (y_embed[:, -1:, :] + eps) * self.scale
-            x_embed = x_embed / (x_embed[:, :, -1:] + eps) * self.scale
-
-        dim_tx = torch.arange(
-            self.num_pos_feats, dtype=torch.float32, device=mask.device)
-        dim_tx = self.temperatureW**(2 * (dim_tx // 2) / self.num_pos_feats)
-        pos_x = x_embed[:, :, :, None] / dim_tx
-
-        dim_ty = torch.arange(
-            self.num_pos_feats, dtype=torch.float32, device=mask.device)
-        dim_ty = self.temperatureH**(2 * (dim_ty // 2) / self.num_pos_feats)
-        pos_y = y_embed[:, :, :, None] / dim_ty
-
-        pos_x = torch.stack(
-            (pos_x[:, :, :, 0::2].sin(), pos_x[:, :, :, 1::2].cos()),
-            dim=4).flatten(3)
-        pos_y = torch.stack(
-            (pos_y[:, :, :, 0::2].sin(), pos_y[:, :, :, 1::2].cos()),
-            dim=4).flatten(3)
-        pos = torch.cat((pos_y, pos_x), dim=3).permute(0, 3, 1, 2)
-
-        # import ipdb; ipdb.set_trace()
-
-        return pos
-
-
-def gen_sineembed_for_position(pos_tensor):
-    # n_query, bs, _ = pos_tensor.size()
-    # sineembed_tensor = torch.zeros(n_query, bs, 256)
-    scale = 2 * math.pi
-    dim_t = torch.arange(128, dtype=torch.float32, device=pos_tensor.device)
-    dim_t = 10000**(2 * (dim_t // 2) / 128)
-    x_embed = pos_tensor[:, :, 0] * scale
-    y_embed = pos_tensor[:, :, 1] * scale
-    pos_x = x_embed[:, :, None] / dim_t
-    pos_y = y_embed[:, :, None] / dim_t
-    pos_x = torch.stack((pos_x[:, :, 0::2].sin(), pos_x[:, :, 1::2].cos()),
-                        dim=3).flatten(2)
-    pos_y = torch.stack((pos_y[:, :, 0::2].sin(), pos_y[:, :, 1::2].cos()),
-                        dim=3).flatten(2)
-    if pos_tensor.size(-1) == 2:
-        pos = torch.cat((pos_y, pos_x), dim=2)
-    elif pos_tensor.size(-1) == 4:
-        w_embed = pos_tensor[:, :, 2] * scale
-        pos_w = w_embed[:, :, None] / dim_t
-        pos_w = torch.stack((pos_w[:, :, 0::2].sin(), pos_w[:, :, 1::2].cos()),
-                            dim=3).flatten(2)
-
-        h_embed = pos_tensor[:, :, 3] * scale
-        pos_h = h_embed[:, :, None] / dim_t
-        pos_h = torch.stack((pos_h[:, :, 0::2].sin(), pos_h[:, :, 1::2].cos()),
-                            dim=3).flatten(2)
-
-        pos = torch.cat((pos_y, pos_x, pos_w, pos_h), dim=2)
-    else:
-        raise ValueError('Unknown pos_tensor shape(-1):{}'.format(
-            pos_tensor.size(-1)))
-    return pos
-
-
-class TransformerDecoder(nn.Module):
-
-    def __init__(
-        self,
-        decoder_layer,
-        num_layers,
-        norm=None,
-        return_intermediate=False,
-        d_model=256,
-        query_dim=2,
-        keep_query_pos=False,
-        query_scale_type='cond_elewise',
-        modulate_hw_attn=False,
-        bbox_embed_diff_each_layer=False,
-    ):
-        super().__init__()
-        self.layers = _get_clones(decoder_layer, num_layers)
-        self.num_layers = num_layers
-        self.norm = norm
-        self.return_intermediate = return_intermediate
-        assert return_intermediate
-        self.query_dim = query_dim
-
-        assert query_scale_type in [
-            'cond_elewise', 'cond_scalar', 'fix_elewise'
-        ]
-        self.query_scale_type = query_scale_type
-        if query_scale_type == 'cond_elewise':
-            self.query_scale = MLP(d_model, d_model, d_model, 2)
-        elif query_scale_type == 'cond_scalar':
-            self.query_scale = MLP(d_model, d_model, 1, 2)
-        elif query_scale_type == 'fix_elewise':
-            self.query_scale = nn.Embedding(num_layers, d_model)
-        else:
-            raise NotImplementedError(
-                'Unknown query_scale_type: {}'.format(query_scale_type))
-
-        self.ref_point_head = MLP(query_dim // 2 * d_model, d_model, d_model,
-                                  2)
-
-        self.bbox_embed = None
-        self.d_model = d_model
-        self.modulate_hw_attn = modulate_hw_attn
-        self.bbox_embed_diff_each_layer = bbox_embed_diff_each_layer
-
-        if modulate_hw_attn:
-            self.ref_anchor_head = MLP(d_model, d_model, 2, 2)
-
-        if not keep_query_pos:
-            for layer_id in range(num_layers - 1):
-                self.layers[layer_id + 1].ca_qpos_proj = None
-
-    def forward(
-            self,
-            tgt,
-            memory,
-            tgt_mask: Optional[Tensor] = None,
-            memory_mask: Optional[Tensor] = None,
-            tgt_key_padding_mask: Optional[Tensor] = None,
-            memory_key_padding_mask: Optional[Tensor] = None,
-            pos: Optional[Tensor] = None,
-            refpoints_unsigmoid: Optional[Tensor] = None,  # num_queries, bs, 2
-    ):
-        output = tgt
-
-        intermediate = []
-        reference_points = refpoints_unsigmoid.sigmoid()
-        ref_points = [reference_points]
-
-        # import ipdb; ipdb.set_trace()
-
-        for layer_id, layer in enumerate(self.layers):
-            obj_center = reference_points[
-                ..., :self.query_dim]  # [num_queries, batch_size, 2]
-            # get sine embedding for the query vector
-            query_sine_embed = gen_sineembed_for_position(obj_center)
-            query_pos = self.ref_point_head(query_sine_embed)
-
-            # For the first decoder layer, we do not apply transformation over p_s
-            if self.query_scale_type != 'fix_elewise':
-                if layer_id == 0:
-                    pos_transformation = 1
-                else:
-                    pos_transformation = self.query_scale(output)
-            else:
-                pos_transformation = self.query_scale.weight[layer_id]
-
-            # apply transformation
-            query_sine_embed = query_sine_embed[
-                ..., :self.d_model] * pos_transformation
-
-            # modulated HW attentions
-            if self.modulate_hw_attn:
-                refHW_cond = self.ref_anchor_head(
-                    output).sigmoid()  # nq, bs, 2
-                query_sine_embed[..., self.d_model //
-                                 2:] *= (refHW_cond[..., 0] /
-                                         obj_center[..., 2]).unsqueeze(-1)
-                query_sine_embed[..., :self.d_model //
-                                 2] *= (refHW_cond[..., 1] /
-                                        obj_center[..., 3]).unsqueeze(-1)
-
-            output = layer(
-                output,
-                memory,
-                tgt_mask=tgt_mask,
-                memory_mask=memory_mask,
-                tgt_key_padding_mask=tgt_key_padding_mask,
-                memory_key_padding_mask=memory_key_padding_mask,
-                pos=pos,
-                query_pos=query_pos,
-                query_sine_embed=query_sine_embed,
-                is_first=(layer_id == 0))
-
-            # iter update
-            if self.bbox_embed is not None:
-                if self.bbox_embed_diff_each_layer:
-                    tmp = self.bbox_embed[layer_id](output)
-                else:
-                    tmp = self.bbox_embed(output)
-                # import ipdb; ipdb.set_trace()
-                tmp[..., :self.query_dim] += inverse_sigmoid(reference_points)
-                new_reference_points = tmp[..., :self.query_dim].sigmoid()
-                if layer_id != self.num_layers - 1:
-                    ref_points.append(new_reference_points)
-                reference_points = new_reference_points.detach()
-
-            if self.return_intermediate:
-                intermediate.append(self.norm(output))
-
-        if self.norm is not None:
-            output = self.norm(output)
-            if self.return_intermediate:
-                intermediate.pop()
-                intermediate.append(output)
-
-        if self.return_intermediate:
-            if self.bbox_embed is not None:
-                return [
-                    torch.stack(intermediate).transpose(1, 2),
-                    torch.stack(ref_points).transpose(1, 2),
-                ]
-            else:
-                return [
-                    torch.stack(intermediate).transpose(1, 2),
-                    reference_points.unsqueeze(0).transpose(1, 2)
-                ]
-
-        return output.unsqueeze(0)
-
-
-class TransformerDecoderLayer(nn.Module):
-
-    def __init__(self,
-                 d_model,
-                 nhead,
-                 dim_feedforward=2048,
-                 dropout=0.1,
-                 activation='relu',
-                 normalize_before=False,
-                 keep_query_pos=False,
-                 rm_self_attn_decoder=False):
-        super().__init__()
-        # Decoder Self-Attention
-        if not rm_self_attn_decoder:
-            self.sa_qcontent_proj = nn.Linear(d_model, d_model)
-            self.sa_qpos_proj = nn.Linear(d_model, d_model)
-            self.sa_kcontent_proj = nn.Linear(d_model, d_model)
-            self.sa_kpos_proj = nn.Linear(d_model, d_model)
-            self.sa_v_proj = nn.Linear(d_model, d_model)
-            self.self_attn = MultiheadAttention(
-                d_model, nhead, dropout=dropout, vdim=d_model)
-
-            self.norm1 = nn.LayerNorm(d_model)
-            self.dropout1 = nn.Dropout(dropout)
-
-        # Decoder Cross-Attention
-        self.ca_qcontent_proj = nn.Linear(d_model, d_model)
-        self.ca_qpos_proj = nn.Linear(d_model, d_model)
-        self.ca_kcontent_proj = nn.Linear(d_model, d_model)
-        self.ca_kpos_proj = nn.Linear(d_model, d_model)
-        self.ca_v_proj = nn.Linear(d_model, d_model)
-        self.ca_qpos_sine_proj = nn.Linear(d_model, d_model)
-        self.cross_attn = MultiheadAttention(
-            d_model * 2, nhead, dropout=dropout, vdim=d_model)
-
-        self.nhead = nhead
-        self.rm_self_attn_decoder = rm_self_attn_decoder
-
-        # Implementation of Feedforward model
-        self.linear1 = nn.Linear(d_model, dim_feedforward)
-        self.dropout = nn.Dropout(dropout)
-        self.linear2 = nn.Linear(dim_feedforward, d_model)
-
-        self.norm2 = nn.LayerNorm(d_model)
-        self.norm3 = nn.LayerNorm(d_model)
-        self.dropout2 = nn.Dropout(dropout)
-        self.dropout3 = nn.Dropout(dropout)
-
-        self.activation = _get_activation_fn(activation)
-        self.normalize_before = normalize_before
-        self.keep_query_pos = keep_query_pos
-
-    def with_pos_embed(self, tensor, pos: Optional[Tensor]):
-        return tensor if pos is None else tensor + pos
-
-    def forward(self,
-                tgt,
-                memory,
-                tgt_mask: Optional[Tensor] = None,
-                memory_mask: Optional[Tensor] = None,
-                tgt_key_padding_mask: Optional[Tensor] = None,
-                memory_key_padding_mask: Optional[Tensor] = None,
-                pos: Optional[Tensor] = None,
-                query_pos: Optional[Tensor] = None,
-                query_sine_embed=None,
-                is_first=False):
-
-        # ========== Begin of Self-Attention =============
-        if not self.rm_self_attn_decoder:
-            # Apply projections here
-            # shape: num_queries x batch_size x 256
-            q_content = self.sa_qcontent_proj(
-                tgt
-            )  # target is the input of the first decoder layer. zero by default.
-            q_pos = self.sa_qpos_proj(query_pos)
-            k_content = self.sa_kcontent_proj(tgt)
-            k_pos = self.sa_kpos_proj(query_pos)
-            v = self.sa_v_proj(tgt)
-
-            num_queries, bs, n_model = q_content.shape
-            hw, _, _ = k_content.shape
-
-            q = q_content + q_pos
-            k = k_content + k_pos
-
-            tgt2 = self.self_attn(
-                q,
-                k,
-                value=v,
-                attn_mask=tgt_mask,
-                key_padding_mask=tgt_key_padding_mask)[0]
-            # ========== End of Self-Attention =============
-
-            tgt = tgt + self.dropout1(tgt2)
-            tgt = self.norm1(tgt)
-
-        # ========== Begin of Cross-Attention =============
-        # Apply projections here
-        # shape: num_queries x batch_size x 256
-        q_content = self.ca_qcontent_proj(tgt)
-        k_content = self.ca_kcontent_proj(memory)
-        v = self.ca_v_proj(memory)
-
-        num_queries, bs, n_model = q_content.shape
-        hw, _, _ = k_content.shape
-
-        k_pos = self.ca_kpos_proj(pos)
-
-        # For the first decoder layer, we concatenate the positional embedding predicted from
-        # the object query (the positional embedding) into the original query (key) in DETR.
-        if is_first or self.keep_query_pos:
-            q_pos = self.ca_qpos_proj(query_pos)
-            q = q_content + q_pos
-            k = k_content + k_pos
-        else:
-            q = q_content
-            k = k_content
-
-        q = q.view(num_queries, bs, self.nhead, n_model // self.nhead)
-        query_sine_embed = self.ca_qpos_sine_proj(query_sine_embed)
-        query_sine_embed = query_sine_embed.view(num_queries, bs, self.nhead,
-                                                 n_model // self.nhead)
-        q = torch.cat([q, query_sine_embed],
-                      dim=3).view(num_queries, bs, n_model * 2)
-        k = k.view(hw, bs, self.nhead, n_model // self.nhead)
-        k_pos = k_pos.view(hw, bs, self.nhead, n_model // self.nhead)
-        k = torch.cat([k, k_pos], dim=3).view(hw, bs, n_model * 2)
-
-        tgt2 = self.cross_attn(
-            query=q,
-            key=k,
-            value=v,
-            attn_mask=memory_mask,
-            key_padding_mask=memory_key_padding_mask)[0]
-        # ========== End of Cross-Attention =============
-
-        tgt = tgt + self.dropout2(tgt2)
-        tgt = self.norm2(tgt)
-        tgt2 = self.linear2(self.dropout(self.activation(self.linear1(tgt))))
-        tgt = tgt + self.dropout3(tgt2)
-        tgt = self.norm3(tgt)
-        return tgt
