+++ easycv/models/detection/detectors/dab_detr/dab_detr_head.py
@@ -8,14 +8,11 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from easycv.models.builder import HEADS, build_neck
-from easycv.models.detection.utils import (HungarianMatcher, accuracy,
+from easycv.models.detection.utils import (HungarianMatcher, SetCriterion,
                                            box_cxcywh_to_xyxy,
-                                           box_xyxy_to_cxcywh,
-                                           generalized_box_iou,
-                                           inverse_sigmoid)
-from easycv.models.loss.focal_loss import py_sigmoid_focal_loss
-from easycv.models.utils import (MLP, get_world_size,
-                                 is_dist_avail_and_initialized)
+                                           box_xyxy_to_cxcywh, inverse_sigmoid)
+from easycv.models.utils import MLP
+from .dn_components import dn_post_process, prepare_for_dn
 
 
 @HEADS.register_module()
@@ -27,15 +24,17 @@ class DABDETRHead(nn.Module):
         num_classes (int): Number of categories excluding the background.
     """
 
-    _version = 2
-
     def __init__(self,
                  num_classes,
                  embed_dims,
                  query_dim=4,
                  iter_update=True,
+                 num_queries=300,
                  num_select=300,
+                 random_refpoints_xy=False,
+                 num_patterns=0,
                  bbox_embed_diff_each_layer=False,
+                 dn_components=None,
                  transformer=None,
                  cost_dict={
                      'cost_class': 1,
@@ -57,7 +56,9 @@ class DABDETRHead(nn.Module):
             num_classes,
             matcher=self.matcher,
             weight_dict=weight_dict,
-            losses=['labels', 'boxes', 'cardinality'])
+            losses=['labels', 'boxes'],
+            loss_class_type='focal_loss',
+            dn_components=dn_components)
         self.postprocess = PostProcess(num_select=num_select)
         self.transformer = build_neck(transformer)
 
@@ -71,8 +72,31 @@ class DABDETRHead(nn.Module):
             self.transformer.decoder.bbox_embed = self.bbox_embed
 
         self.num_classes = num_classes
+        self.num_queries = num_queries
+        self.embed_dims = embed_dims
         self.query_dim = query_dim
         self.bbox_embed_diff_each_layer = bbox_embed_diff_each_layer
+        self.dn_components = dn_components
+
+        self.query_embed = nn.Embedding(num_queries, query_dim)
+        self.random_refpoints_xy = random_refpoints_xy
+        if random_refpoints_xy:
+            self.query_embed.weight.data[:, :2].uniform_(0, 1)
+            self.query_embed.weight.data[:, :2] = inverse_sigmoid(
+                self.query_embed.weight.data[:, :2])
+            self.query_embed.weight.data[:, :2].requires_grad = False
+
+        self.num_patterns = num_patterns
+        if not isinstance(num_patterns, int):
+            Warning('num_patterns should be int but {}'.format(
+                type(num_patterns)))
+            self.num_patterns = 0
+        if self.num_patterns > 0:
+            self.patterns = nn.Embedding(self.num_patterns, embed_dims)
+
+        if self.dn_components:
+            # leave one dim for indicator
+            self.label_enc = nn.Embedding(num_classes + 1, embed_dims - 1)
 
     def init_weights(self):
         self.transformer.init_weights()
@@ -92,7 +116,45 @@ class DABDETRHead(nn.Module):
             nn.init.constant_(self.bbox_embed.layers[-1].weight.data, 0)
             nn.init.constant_(self.bbox_embed.layers[-1].bias.data, 0)
 
-    def forward(self, feats, img_metas):
+    def prepare(self, feats, targets=None, mode='train'):
+        bs = feats[0].shape[0]
+        query_embed = self.query_embed.weight
+        if self.dn_components:
+            # default pipeline
+            self.dn_components['num_patterns'] = self.num_patterns
+            self.dn_components['targets'] = targets
+            # prepare for dn
+            tgt, query_embed, attn_mask, mask_dict = prepare_for_dn(
+                mode, self.dn_components, query_embed, bs, self.num_queries,
+                self.num_classes, self.embed_dims, self.label_enc)
+            if self.num_patterns > 0:
+                l = tgt.shape[0]
+                tgt[l - self.num_queries * self.num_patterns:] += \
+                    self.patterns.weight[:, None, None, :].repeat(1, self.num_queries, bs, 1).flatten(0, 1)
+            return query_embed, tgt, attn_mask, mask_dict
+        else:
+            query_embed = query_embed.unsqueeze(1).repeat(1, bs, 1)
+            if self.num_patterns == 0:
+                tgt = torch.zeros(
+                    self.num_queries,
+                    bs,
+                    self.embed_dims,
+                    device=query_embed.device)
+            else:
+                tgt = self.patterns.weight[:, None, None, :].repeat(
+                    1, self.num_queries, bs,
+                    1).flatten(0, 1)  # n_q*n_pat, bs, d_model
+                query_embed = query_embed.repeat(self.num_patterns, 1,
+                                                 1)  # n_q*n_pat, bs, d_model
+        return query_embed, tgt, None, None
+
+    def forward(self,
+                feats,
+                img_metas,
+                query_embed=None,
+                tgt=None,
+                attn_mask=None,
+                mask_dict=None):
         """Forward function.
         Args:
             feats (tuple[Tensor]): Features from the upstream network, each is
@@ -109,7 +171,9 @@ class DABDETRHead(nn.Module):
                     normalized coordinate format (cx, cy, w, h) and shape \
                     [nb_dec, bs, num_query, 4].
         """
-        feats = self.transformer(feats, img_metas)
+
+        feats = self.transformer(
+            feats, img_metas, query_embed, tgt, attn_mask=attn_mask)
 
         hs, reference = feats
         outputs_class = self.class_embed(hs)
@@ -128,6 +192,10 @@ class DABDETRHead(nn.Module):
                 outputs_coords.append(outputs_coord)
             outputs_coord = torch.stack(outputs_coords)
 
+        if mask_dict is not None:
+            # dn post process
+            outputs_class, outputs_coord = dn_post_process(
+                outputs_class, outputs_coord, mask_dict)
         out = {
             'pred_logits': outputs_class[-1],
             'pred_boxes': outputs_coord[-1]
@@ -164,27 +232,45 @@ class DABDETRHead(nn.Module):
         Returns:
             dict[str, Tensor]: A dictionary of loss components.
         """
-        outputs = self.forward(x, img_metas)
-
+        # prepare ground truth
         for i in range(len(img_metas)):
             img_h, img_w, _ = img_metas[i]['img_shape']
             # DETR regress the relative position of boxes (cxcywh) in the image.
             # Thus the learning target should be normalized by the image size, also
             # the box format should be converted from defaultly x1y1x2y2 to cxcywh.
-            factor = outputs['pred_boxes'].new_tensor(
-                [img_w, img_h, img_w, img_h]).unsqueeze(0)
+            factor = gt_bboxes[i].new_tensor([img_w, img_h, img_w,
+                                              img_h]).unsqueeze(0)
             gt_bboxes[i] = box_xyxy_to_cxcywh(gt_bboxes[i]) / factor
 
         targets = []
         for gt_label, gt_bbox in zip(gt_labels, gt_bboxes):
             targets.append({'labels': gt_label, 'boxes': gt_bbox})
 
-        losses = self.criterion(outputs, targets)
+        query_embed, tgt, attn_mask, mask_dict = self.prepare(
+            x, targets=targets, mode='train')
+
+        outputs = self.forward(
+            x,
+            img_metas,
+            query_embed=query_embed,
+            tgt=tgt,
+            attn_mask=attn_mask,
+            mask_dict=mask_dict)
+
+        losses = self.criterion(outputs, targets, mask_dict)
 
         return losses
 
     def forward_test(self, x, img_metas):
-        outputs = self.forward(x, img_metas)
+        query_embed, tgt, attn_mask, mask_dict = self.prepare(x, mode='test')
+
+        outputs = self.forward(
+            x,
+            img_metas,
+            query_embed=query_embed,
+            tgt=tgt,
+            attn_mask=attn_mask,
+            mask_dict=mask_dict)
 
         ori_shape_list = []
         for i in range(len(img_metas)):
@@ -242,192 +328,3 @@ class PostProcess(nn.Module):
         }
 
         return results
-
-
-class SetCriterion(nn.Module):
-    """ This class computes the loss for Conditional DETR.
-    The process happens in two steps:
-        1) we compute hungarian assignment between ground truth boxes and the outputs of the model
-        2) we supervise each pair of matched ground-truth / prediction (supervise class and box)
-    """
-
-    def __init__(self, num_classes, matcher, weight_dict, losses):
-        """ Create the criterion.
-        Parameters:
-            num_classes: number of object categories, omitting the special no-object category
-            matcher: module able to compute a matching between targets and proposals
-            weight_dict: dict containing as key the names of the losses and as values their relative weight.
-            losses: list of all the losses to be applied. See get_loss for list of available losses.
-        """
-        super().__init__()
-        self.num_classes = num_classes
-        self.matcher = matcher
-        self.weight_dict = weight_dict
-        self.losses = losses
-
-    def loss_labels(self, outputs, targets, indices, num_boxes, log=True):
-        """Classification loss (Binary focal loss)
-        targets dicts must contain the key "labels" containing a tensor of dim [nb_target_boxes]
-        """
-        assert 'pred_logits' in outputs
-        src_logits = outputs['pred_logits']
-
-        idx = self._get_src_permutation_idx(indices)
-        target_classes_o = torch.cat(
-            [t['labels'][J] for t, (_, J) in zip(targets, indices)])
-        target_classes = torch.full(
-            src_logits.shape[:2],
-            self.num_classes,
-            dtype=torch.int64,
-            device=src_logits.device)
-        target_classes[idx] = target_classes_o
-
-        target_classes_onehot = torch.zeros([
-            src_logits.shape[0], src_logits.shape[1], src_logits.shape[2] + 1
-        ],
-                                            dtype=src_logits.dtype,
-                                            layout=src_logits.layout,
-                                            device=src_logits.device)
-        target_classes_onehot.scatter_(2, target_classes.unsqueeze(-1), 1)
-
-        target_classes_onehot = target_classes_onehot[:, :, :-1]
-        loss_ce = py_sigmoid_focal_loss(
-            src_logits,
-            target_classes_onehot.long(),
-            alpha=0.25,
-            gamma=2,
-            reduction='none').mean(1).sum() / num_boxes
-        loss_ce = loss_ce * src_logits.shape[1] * self.weight_dict['loss_ce']
-        losses = {'loss_ce': loss_ce}
-
-        if log:
-            # TODO this should probably be a separate loss, not hacked in this one here
-            losses['class_error'] = 100 - accuracy(src_logits[idx],
-                                                   target_classes_o)[0]
-        return losses
-
-    @torch.no_grad()
-    def loss_cardinality(self, outputs, targets, indices, num_boxes):
-        """ Compute the cardinality error, ie the absolute error in the number of predicted non-empty boxes
-        This is not really a loss, it is intended for logging purposes only. It doesn't propagate gradients
-        """
-        pred_logits = outputs['pred_logits']
-        device = pred_logits.device
-        tgt_lengths = torch.as_tensor([len(v['labels']) for v in targets],
-                                      device=device)
-        # Count the number of predictions that are NOT "no-object" (which is the last class)
-        card_pred = (pred_logits.argmax(-1) !=
-                     pred_logits.shape[-1] - 1).sum(1)
-        card_err = F.l1_loss(card_pred.float(), tgt_lengths.float())
-        losses = {'cardinality_error': card_err}
-        return losses
-
-    def loss_boxes(self, outputs, targets, indices, num_boxes):
-        """Compute the losses related to the bounding boxes, the L1 regression loss and the GIoU loss
-           targets dicts must contain the key "boxes" containing a tensor of dim [nb_target_boxes, 4]
-           The target boxes are expected in format (center_x, center_y, w, h), normalized by the image size.
-        """
-        assert 'pred_boxes' in outputs
-        idx = self._get_src_permutation_idx(indices)
-        src_boxes = outputs['pred_boxes'][idx]
-        target_boxes = torch.cat(
-            [t['boxes'][i] for t, (_, i) in zip(targets, indices)], dim=0)
-
-        loss_bbox = F.l1_loss(src_boxes, target_boxes, reduction='none')
-
-        losses = {}
-        losses['loss_bbox'] = loss_bbox.sum(
-        ) / num_boxes * self.weight_dict['loss_bbox']
-
-        loss_giou = 1 - torch.diag(
-            generalized_box_iou(
-                box_cxcywh_to_xyxy(src_boxes),
-                box_cxcywh_to_xyxy(target_boxes)))
-        losses['loss_giou'] = loss_giou.sum(
-        ) / num_boxes * self.weight_dict['loss_giou']
-
-        return losses
-
-    def _get_src_permutation_idx(self, indices):
-        # permute predictions following indices
-        batch_idx = torch.cat(
-            [torch.full_like(src, i) for i, (src, _) in enumerate(indices)])
-        src_idx = torch.cat([src for (src, _) in indices])
-        return batch_idx, src_idx
-
-    def _get_tgt_permutation_idx(self, indices):
-        # permute targets following indices
-        batch_idx = torch.cat(
-            [torch.full_like(tgt, i) for i, (_, tgt) in enumerate(indices)])
-        tgt_idx = torch.cat([tgt for (_, tgt) in indices])
-        return batch_idx, tgt_idx
-
-    def get_loss(self, loss, outputs, targets, indices, num_boxes, **kwargs):
-        loss_map = {
-            'labels': self.loss_labels,
-            'cardinality': self.loss_cardinality,
-            'boxes': self.loss_boxes,
-        }
-        assert loss in loss_map, f'do you really want to compute {loss} loss?'
-        return loss_map[loss](outputs, targets, indices, num_boxes, **kwargs)
-
-    def forward(self, outputs, targets, return_indices=False):
-        """ This performs the loss computation.
-        Parameters:
-             outputs: dict of tensors, see the output specification of the model for the format
-             targets: list of dicts, such that len(targets) == batch_size.
-                      The expected keys in each dict depends on the losses applied, see each loss' doc
-
-             return_indices: used for vis. if True, the layer0-5 indices will be returned as well.
-        """
-
-        outputs_without_aux = {
-            k: v
-            for k, v in outputs.items() if k != 'aux_outputs'
-        }
-
-        # Retrieve the matching between the outputs of the last layer and the targets
-        indices = self.matcher(outputs_without_aux, targets)
-        if return_indices:
-            indices0_copy = indices
-            indices_list = []
-
-        # Compute the average number of target boxes accross all nodes, for normalization purposes
-        num_boxes = sum(len(t['labels']) for t in targets)
-        num_boxes = torch.as_tensor([num_boxes],
-                                    dtype=torch.float,
-                                    device=next(iter(outputs.values())).device)
-        if is_dist_avail_and_initialized():
-            torch.distributed.all_reduce(num_boxes)
-        num_boxes = torch.clamp(num_boxes / get_world_size(), min=1).item()
-
-        # Compute all the requested losses
-        losses = {}
-        for loss in self.losses:
-            losses.update(
-                self.get_loss(loss, outputs, targets, indices, num_boxes))
-
-        # In case of auxiliary losses, we repeat this process with the output of each intermediate layer.
-        if 'aux_outputs' in outputs:
-            for i, aux_outputs in enumerate(outputs['aux_outputs']):
-                indices = self.matcher(aux_outputs, targets)
-                if return_indices:
-                    indices_list.append(indices)
-                for loss in self.losses:
-                    if loss == 'masks':
-                        # Intermediate masks losses are too costly to compute, we ignore them.
-                        continue
-                    kwargs = {}
-                    if loss == 'labels':
-                        # Logging is enabled only for the last layer
-                        kwargs = {'log': False}
-                    l_dict = self.get_loss(loss, aux_outputs, targets, indices,
-                                           num_boxes, **kwargs)
-                    l_dict = {k + f'_{i}': v for k, v in l_dict.items()}
-                    losses.update(l_dict)
-
-        if return_indices:
-            indices_list.append(indices0_copy)
-            return losses, indices_list
-
-        return losses
