+++ easycv/models/detection/detectors/dab_detr/dn_components.py
@@ -170,163 +170,3 @@ def dn_post_process(outputs_class, outputs_coord, mask_dict):
         mask_dict['output_known_lbs_bboxes'] = (output_known_class,
                                                 output_known_coord)
     return outputs_class, outputs_coord
-
-
-class DNCriterion(nn.Module):
-    """ This class computes the loss for Conditional DETR.
-    The process happens in two steps:
-        1) we compute hungarian assignment between ground truth boxes and the outputs of the model
-        2) we supervise each pair of matched ground-truth / prediction (supervise class and box)
-    """
-
-    def __init__(self, weight_dict):
-        """ Create the criterion.
-        Parameters:
-            num_classes: number of object categories, omitting the special no-object category
-            matcher: module able to compute a matching between targets and proposals
-            weight_dict: dict containing as key the names of the losses and as values their relative weight.
-            losses: list of all the losses to be applied. See get_loss for list of available losses.
-        """
-        super().__init__()
-        self.weight_dict = weight_dict
-
-    def prepare_for_loss(self, mask_dict):
-        """
-        prepare dn components to calculate loss
-        Args:
-            mask_dict: a dict that contains dn information
-        """
-        output_known_class, output_known_coord = mask_dict[
-            'output_known_lbs_bboxes']
-        known_labels, known_bboxs = mask_dict['known_lbs_bboxes']
-        map_known_indice = mask_dict['map_known_indice']
-
-        known_indice = mask_dict['known_indice']
-
-        batch_idx = mask_dict['batch_idx']
-        bid = batch_idx[known_indice]
-        if len(output_known_class) > 0:
-            output_known_class = output_known_class.permute(
-                1, 2, 0, 3)[(bid, map_known_indice)].permute(1, 0, 2)
-            output_known_coord = output_known_coord.permute(
-                1, 2, 0, 3)[(bid, map_known_indice)].permute(1, 0, 2)
-        num_tgt = known_indice.numel()
-        return known_labels, known_bboxs, output_known_class, output_known_coord, num_tgt
-
-    def tgt_loss_boxes(
-        self,
-        src_boxes,
-        tgt_boxes,
-        num_tgt,
-    ):
-        """Compute the losses related to the bounding boxes, the L1 regression loss and the GIoU loss
-        targets dicts must contain the key "boxes" containing a tensor of dim [nb_target_boxes, 4]
-        The target boxes are expected in format (center_x, center_y, w, h), normalized by the image size.
-        """
-        if len(tgt_boxes) == 0:
-            return {
-                'tgt_loss_bbox': torch.as_tensor(0.).to('cuda'),
-                'tgt_loss_giou': torch.as_tensor(0.).to('cuda'),
-            }
-
-        loss_bbox = F.l1_loss(src_boxes, tgt_boxes, reduction='none')
-
-        losses = {}
-        losses['tgt_loss_bbox'] = loss_bbox.sum(
-        ) / num_tgt * self.weight_dict['loss_bbox']
-
-        loss_giou = 1 - torch.diag(
-            generalized_box_iou(
-                box_cxcywh_to_xyxy(src_boxes), box_cxcywh_to_xyxy(tgt_boxes)))
-        losses['tgt_loss_giou'] = loss_giou.sum(
-        ) / num_tgt * self.weight_dict['loss_giou']
-        return losses
-
-    def tgt_loss_labels(self,
-                        src_logits_,
-                        tgt_labels_,
-                        num_tgt,
-                        focal_alpha,
-                        log=False):
-        """Classification loss (NLL)
-        targets dicts must contain the key "labels" containing a tensor of dim [nb_target_boxes]
-        """
-        if len(tgt_labels_) == 0:
-            return {
-                'tgt_loss_ce': torch.as_tensor(0.).to('cuda'),
-                'tgt_class_error': torch.as_tensor(0.).to('cuda'),
-            }
-
-        src_logits, tgt_labels = src_logits_.unsqueeze(
-            0), tgt_labels_.unsqueeze(0)
-
-        target_classes_onehot = torch.zeros([
-            src_logits.shape[0], src_logits.shape[1], src_logits.shape[2] + 1
-        ],
-                                            dtype=src_logits.dtype,
-                                            layout=src_logits.layout,
-                                            device=src_logits.device)
-        target_classes_onehot.scatter_(2, tgt_labels.unsqueeze(-1), 1)
-
-        target_classes_onehot = target_classes_onehot[:, :, :-1]
-        loss_ce = py_sigmoid_focal_loss(
-            src_logits,
-            target_classes_onehot.long(),
-            alpha=focal_alpha,
-            gamma=2,
-            reduction='none').mean(1).sum(
-            ) / num_tgt * src_logits.shape[1] * self.weight_dict['loss_ce']
-
-        losses = {'tgt_loss_ce': loss_ce}
-        if log:
-            losses['tgt_class_error'] = 100 - accuracy(src_logits_,
-                                                       tgt_labels_)[0]
-        return losses
-
-    def forward(self, mask_dict, training, aux_num, focal_alpha):
-        """
-        compute dn loss in criterion
-        Args:
-            mask_dict: a dict for dn information
-            training: training or inference flag
-            aux_num: aux loss number
-            focal_alpha:  for focal loss
-        """
-        losses = {}
-        if training and 'output_known_lbs_bboxes' in mask_dict:
-            known_labels, known_bboxs, output_known_class, output_known_coord, num_tgt = self.prepare_for_loss(
-                mask_dict)
-            losses.update(
-                self.tgt_loss_labels(output_known_class[-1], known_labels,
-                                     num_tgt, focal_alpha))
-            losses.update(
-                self.tgt_loss_boxes(output_known_coord[-1], known_bboxs,
-                                    num_tgt))
-        else:
-            losses['tgt_loss_bbox'] = torch.as_tensor(0.).to('cuda')
-            losses['tgt_loss_giou'] = torch.as_tensor(0.).to('cuda')
-            losses['tgt_loss_ce'] = torch.as_tensor(0.).to('cuda')
-            losses['tgt_class_error'] = torch.as_tensor(0.).to('cuda')
-
-        if aux_num:
-            for i in range(aux_num):
-                # dn aux loss
-                if training and 'output_known_lbs_bboxes' in mask_dict:
-                    l_dict = self.tgt_loss_labels(output_known_class[i],
-                                                  known_labels, num_tgt,
-                                                  focal_alpha)
-                    l_dict = {k + f'_{i}': v for k, v in l_dict.items()}
-                    losses.update(l_dict)
-                    l_dict = self.tgt_loss_boxes(output_known_coord[i],
-                                                 known_bboxs, num_tgt)
-                    l_dict = {k + f'_{i}': v for k, v in l_dict.items()}
-                    losses.update(l_dict)
-                else:
-                    l_dict = dict()
-                    l_dict['tgt_loss_bbox'] = torch.as_tensor(0.).to('cuda')
-                    l_dict['tgt_class_error'] = torch.as_tensor(0.).to('cuda')
-                    l_dict['tgt_loss_giou'] = torch.as_tensor(0.).to('cuda')
-                    l_dict['tgt_loss_ce'] = torch.as_tensor(0.).to('cuda')
-                    l_dict = {k + f'_{i}': v for k, v in l_dict.items()}
-                    losses.update(l_dict)
-        return losses
