--- easycv/models/detection/detr/__init__.py
@@ -1,2 +1,0 @@
-from .detr_head import DETRHead
-from .detr_transformer import DetrTransformer
